Move some lonely libs from lib/ to app/lib
This commit is one of some to move mostly of our lib/* to app/lib
in order to have eager loading on boot

This one moves some lonely libs from lib/ to app/lib/.
