+++ app/lib/cross_domain_sessions.rb
