+++ app/lib/simple_layout.rb
