+++ test/unit/abilities/accounts_test.rb
@@ -1,12 +1,14 @@
-require File.expand_path(File.dirname(__FILE__) + '/../../test_helper')
+require 'test_helper'
 
-class Abilities::AccountsTest < ActiveSupport::TestCase
-  def setup
-    @provider = FactoryBot.create(:provider_account)
-  end
+module Abilities
+  class AccountsTest < ActiveSupport::TestCase
+    def setup
+      @provider = FactoryBot.create(:provider_account)
+    end
 
-  test "admin can't destroy himself" do
-    user = @provider.admins.first
-    assert_cannot Ability.new(user), :destroy, user
+    test "admin can't destroy himself" do
+      user = @provider.admins.first
+      assert_cannot Ability.new(user), :destroy, user
+    end
   end
 end
